Stop importing testing into Virtlet
This was causing lots of testing flags to appear in virtletctl help
